Add chat integration to event-application-pair and account
In this commit, we've added attributes related to chats to the 'event-application-pair' and 'account' types. 'ChatId' is added to event-application-pair type to allow linking an event application pair to a chat. On 'account' type, 'chatMembers' and 'messages' are added to manage chat members and messages for an account. These changes support the feature of chat integration into the application, providing improved communication features.