"Refactor group service methods to use accountId instead of account object
In the group service, methods were refactored to pass accountId as a parameter instead of the whole account object. This was done to make the function signatures more clear and intuitive and ease the method's usage. Additional relevant changes were made in group resolvers to keep consistency across the application."