Implement forwardRef for handling circular dependencies
In order to manage circular dependencies between Account, Group, and AccountSession modules, forwardRef() has been used. This helps NestJS to load classes of possibly circular dependent modules in a deferred way. In addition, the ProfileService is now being exported from the ProfileModule, eliminating the need to provide it directly in the AccountModule. This is in line with the latest updates to the NestJS framework, which promotes better modularity and code cohesiveness.