Added methods to fetch events and group invites by group ID
This commit adds two important functionalities to the group module. It introduces the ability to fetch events and group invites associated with a specific group ID. Specifically, 'getEventsByGroupId' method in 'eventService' and 'getGroupInvitesByGroupId' method in 'groupInviteService' are introduced. Adequate resolvers are also added in 'groupResolver'.  The Event Module is also modified to export the EventService, hence making it available to GroupModule. Additionally, new imports are included in the EventModule and GroupModule to ensure proper functioning of these enhancements. Modifications are also done on 'EventResolver' to facilitate fetching of group details and managing events.