Update group fetching and group image default URL
Modified group fetching to ensure limits to the accountId in the Prisma 'groupMember' findMany method. This ensures that only groups where the user is a member are returned in the response, adding extra security.

Additionally, the group's default image URL was updated for more a context-specific image.

In group.resolver.ts, 'offset' and 'limit' in the methods 'publicGroups' and 'privateGroups' were typed as Int to ensure that only integer values are passed as arguments, promoting type safety.