Remove unnecessary getGroupId method and add 'ADMIN' role check
The 'getGroupId' method in group.service.ts was removed as it was redundant code and not being used in the application.

Changed the conditional check in the 'ifAccountAdminOfGroup' method to also consider if a member has an 'ADMIN' role. This ensures users who are group administrators can perform certain actions restricted to admin level.

Also, in 'createGroup' method, explicitly assigning 'ADMIN' role to the user when a new group is created, assigning him the admin privileges.