Refactored group service to filter closed groups & refined user role checks
The changes on group.service.ts added additional filter conditions to skip groups with a 'CLOSED' status when users list 'PUBLIC' or 'PRIVATE' groups.

The 'ifAccountAdminOfGroup' method was renamed to 'isMemberGroupAdmin', fetching whether the member has admin privileges regardless of the group type. This modification simplifies the retrieval process and ensures the user role is correctly identified across both 'PUBLIC' and 'PRIVATE' groups.

For the 'closeGroup' method, the admin check was removed as only admins can close groups. This is done externally via the resolver.

In group.resolver.ts, 'checkIfMemberIsGroupAdmin' was added to ensure that only an admin user can update or delete a group.

The 'event-application-pair.resolver.ts' file was also updated to replace 'ifAccountAdminOfGroup' method with the new 'isMemberGroupAdmin'.