Add GroupInvite mutation and modify group id Args
This commit includes the necessary changes needed for creating group invites on the server side. An 'Int' type option was added to 'Args' decorator of the 'group' function in Foundgroup.resolver.ts to fulfill GraphQL expectations for the type specification.

Additionally, a new 'createGroupInvite' mutation was introduced in group-invite.resolver.ts for creating and returning a new group invite code, for the group with the provided groupId. This mutation function will be utilized to create invite codes for group members.