Ensure query arguments are Int in event resolver
The type of arguments in the event resolver has been updated to improve data integrity in queries. The `groupId` and `id` arguments have been explicitly typed as `Int` to ensure that only numbers are inputted, enhancing data security and reducing potential system errors. Also, the decorator `@Field` has been modified to enforce input as `Int`. This change drastically lowers the risk of type