Add method getGroupById and query groupByEventId
Implemented 'getGroupById' method in group service and 'groupByEventId' query in group resolver to enable fetching of a specific group based on its 'id'. This functionality is essential for retrieving and manipulating group data individually. Also, replaced 'getGroupByEventId' method call with 'getGroupById' in event resolver to properly fetch the related group of an event based on group 'id' not event 'id'.