Update field type and refine methods in app
Changed the types of "accountId", "eventId", and "applicationId" fields to Int in the CreateEventApplicationInput and CreatePreferenceInput classes under event-application and preference resolvers respectively. Updated "getEventApplicationPairById" method in the event-application-pair service to log and return the result for extra debugging information. Also adjusted the way 'query' is handled for "id" in "eventApplicationPair" method and renamed "eventApplicationFirst" to "applicationFirst", "eventApplicationSecond" to "applicationSecond" in event-application-pair resolver. This enhances data precision and readability.