Refactor: Removed unused Int import for preference resolver
Removed the Int descriptor from the CreatePreferenceInput class on the preference resolver, as it was no longer being used following a specification change. Moreover, the "applicationId" field, that was previously defined as 'Int', has been removed from the class. This change simplifies and cleans up the code, improves readability, and enhances maintainability with the most current project specifications.