Add: New method to get event app pair by event and account
A new async method 'getEventApplicationPairByEventAndAccount' was added to event-application-pair service. It's purpose is to find and return the first matching pair for given eventId and accountId using the prisma service. An equivalent was also added to the event-application-pair resolver to handle relevant queries. Unused import 'console' was removed as part of the cleanup. This method expands the way the application handles and retrieves data, giving more flexibility to the service's usage.