Add unique avatar feature and account count query
The update includes avatar generation for each unique externalId when creating a profile. It uses the md5 hash of the externalId and a list of seeds to generate a unique avatar URL from the DiceBear Avatars API. The seeds are just a list of strings. This ensures each user has a unique yet consistent avatar.

Also added is a method to get the total number of accounts in the account service, alongside a corresponding query in the account resolver which guards the method behind authentication.

Finally, updated package-lock.json to include the package for DiceBear Avatar. This library is used for generating SVG avatars.