Make pictureUrl field nullable for Group and Event
The `pictureUrl` field of `Group` and `Event` tables in the database has been made optional since it's not always necessary to create a new group or event with a pictureUrl.
This change has been reflected in the related Prisma schema and GraphQL queries.
All the create and update handlers in the Group and Event services now check for existence of `pictureUrl` before operating on the field.
A hardcoded image url has also been removed in favor of accepting dynamic input.
Unit tests have been updated to accommodate these changes.