Add isNotificationsEnabled to Account model
This commit includes a new attribute 'isNotificationsEnabled' in the Account model. This was done by updating the schema in prisma and consequently updating type definitions and inputs in the generated typescript and nestgraphql files. The new attribute is a boolean value that indicates whether the user has enabled notifications. The SQL migration file for this change was also generated.

The update was necessary to supports user’s preference for notifications and would be used to control the sending of notifications to users. The default value for the isNotificationsEnabled is set as true, considering usability and user engagement.