Replace domain URL with event name in notification messages
This commit removes the usage of `process.env.SECRET_SANTA_DOMAIN` in notification messages and replaces it with the name of the event. This change was made to make the notification messages more friendly and easier to understand, as not all users may be familiar with URLs. The change was implemented across multiple methods in the event-application and event-application-pair services. The process module was also removed as it's no longer in use.